Compilation bugfixes.

+++ Ball.h
@@ -34,7 +34,7 @@ const unsigned int BALL_MASK[DIAMETER][DIAMETER] = {
                                 {NONE,NONE,NONE,NONE,NONE,FILL,FILL,FILL,FILL,FILL,NONE,NONE,NONE,NONE,NONE}
                             };
 
-static const unsigned int SIN_TABLE[360] = {
+static const double SIN_TABLE[360] = {
                                 0.000,  0.017, 0.035, 0.052, 0.070, 0.087, 0.105, 0.122, 0.139, 0.156, 0.174, 0.191, 0.208, 0.225, 0.242, 0.259, 
                                 0.276, 0.292, 0.309, 0.326, 0.342, 0.358, 0.375, 0.391, 0.407, 0.423, 0.438, 0.454, 0.469, 0.485, 0.500, 0.515, 
                                 0.530, 0.545, 0.559, 0.574, 0.588, 0.602, 0.616, 0.629, 0.643, 0.656, 0.669, 0.682, 0.695, 0.707, 0.719, 0.731,
